Add files via upload
This C++ program counts the number of digits in a given integer. It prompts the user to enter a number and then calculates the total number of digits by repeatedly dividing the number by 10 until it reaches zero. This program serves as a simple example of using loops and basic arithmetic operations in C++.