Add files via upload
This C++ program automates the process of abbreviating long words, defined as words with a length strictly greater than 10 characters. The abbreviation is created by taking the first letter, counting the number of letters between the first and last letters, and appending the last letter. This approach simplifies the representation of lengthy words, making them easier to read and manage in text.