Add files via upload
Title: Minimum Steps for the Elephant to Reach His Friend's House

As a first-year student diving into programming, I recently tackled a problem involving an elephant trying to reach his friend's house located at a certain coordinate on a number line. The challenge was to determine the minimum number of steps the elephant needs to take, given that he can move between 1 to 5 positions in a single step.

In my solution, I implemented a simple algorithm in C++ that calculates the minimum steps based on the distance to the friend's house. The logic is straightforward: if the distance (x) is divisible by 5, the elephant can reach his friend's house in exactly $ \frac{x}{5} $ steps. If not, he will need one additional step to cover the remaining distance.