Add files via upload
🌟 Excited to share my latest coding achievement! 🌟

As a first-year B.Tech student specializing in Computer Science Engineering, I’m diving into the world of Data Structures and Algorithms (DSA). Recently, I tackled a problem from Codeforces that challenged me to calculate the minimum width of a road for Vanya and his friends to walk unnoticed by a guard.

The problem required me to consider the heights of each friend and determine their walking width based on whether they needed to bend down or not. If a friend's height exceeds a certain limit, they take up more space while walking. This was a great exercise in understanding conditional logic and loops!