Add files via upload
Title: Exploring Nearly Lucky Numbers in C++

As a first-year student, I recently delved into the concept of nearly lucky numbers in programming. A nearly lucky number is defined as a number that contains a count of lucky digits (4 and 7) that itself is a lucky number. In my latest C++ project, I implemented a program to determine if a given number is nearly lucky.

Through this exercise, I learned about digit manipulation and condition checking in C++. It was a great way to apply my coding skills and deepen my understanding of number theory. I’m excited to share my code and insights with you all!