Add files via upload
Implemented a solution to execute a program written in the Bit++ language, which consists of a single variable $ x $ initialized to 0 and two operations: increment (++) and decrement (--). The program processes a sequence of statements, updating the value of $ x $ based on the operations specified. The final value of $ x $ is computed after executing all statements, demonstrating the functionality of the Bit++ language.